@@ -1,78 +1,78 @@
 {
-  "name":            "@heypoint/source",
-  "version":         "0.0.0",
-  "homepage":        "https://heypoint.app",
-  "license":         "Apache-2.0",
-  "private":         true,
-  "author":          {
-    "name":  "Gavin Sawyer",
+  "name": "@heypoint/source",
+  "version": "0.0.0",
+  "homepage": "https://heypoint.app",
+  "license": "Apache-2.0",
+  "private": true,
+  "author": {
+    "name": "Gavin Sawyer",
     "email": "g@vinsawyer.dev",
-    "url":   "https://gavinsawyer.dev"
+    "url": "https://gavinsawyer.dev"
   },
-  "repository":      {
+  "repository": {
     "type": "git",
-    "url":  "git+https://github.com/heypoint/heypoint.git"
+    "url": "git+https://github.com/heypoint/heypoint.git"
   },
-  "scripts":         {
-    "build":  "npx nx run-many --target build --all",
+  "scripts": {
+    "build": "npx nx run-many --target build --all",
     "deploy": "npx nx run-many --target deploy --all",
-    "lint":   "npx nx run-many --target lint --all"
+    "lint": "npx nx run-many --target lint --all"
   },
-  "dependencies":    {
-    "@angular/animations":               "^16.1.7",
-    "@angular/cdk":                      "^16.1.6",
-    "@angular/common":                   "^16.1.7",
-    "@angular/compiler":                 "^16.1.7",
-    "@angular/core":                     "^16.1.7",
-    "@angular/fire":                     "^7.6.1",
-    "@angular/forms":                    "^16.1.8",
-    "@angular/google-maps":              "^16.1.7",
-    "@angular/material":                 "^16.1.7",
-    "@angular/platform-browser":         "^16.1.7",
-    "@angular/platform-browser-dynamic": "^16.1.8",
-    "@angular/platform-server":          "^16.1.8",
-    "@angular/router":                   "^16.1.8",
-    "@nguniversal/express-engine":       "^16.1.2",
-    "rxjs":                              "^7.8.1",
-    "tslib":                             "^2.6.1",
-    "zone.js":                           "^0.13.1"
+  "dependencies": {
+    "@angular/animations": "^16.1.7",
+    "@angular/cdk": "^16.1.6",
+    "@angular/common": "^16.1.7",
+    "@angular/compiler": "^16.1.7",
+    "@angular/core": "^16.1.7",
+    "@angular/fire": "^7.6.1",
+    "@angular/forms": "^16.2.0",
+    "@angular/google-maps": "^16.2.0",
+    "@angular/material": "^16.2.0",
+    "@angular/platform-browser": "^16.1.7",
+    "@angular/platform-browser-dynamic": "^16.2.0",
+    "@angular/platform-server": "^16.2.0",
+    "@angular/router": "^16.2.0",
+    "@nguniversal/express-engine": "^16.2.0",
+    "rxjs": "^7.8.1",
+    "tslib": "^2.6.1",
+    "zone.js": "^0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular":          "^16.1.8",
-    "@angular-devkit/core":                   "^16.1.6",
-    "@angular-devkit/schematics":             "^16.1.6",
-    "@angular-eslint/eslint-plugin":          "^16.1.0",
+    "@angular-devkit/build-angular": "^16.2.0",
+    "@angular-devkit/core": "^16.1.6",
+    "@angular-devkit/schematics": "^16.1.6",
+    "@angular-eslint/eslint-plugin": "^16.1.0",
     "@angular-eslint/eslint-plugin-template": "^16.1.0",
-    "@angular-eslint/template-parser":        "^16.1.0",
-    "@angular/cli":                           "^16.1.8",
-    "@angular/compiler-cli":                  "^16.1.8",
-    "@angular/language-service":              "^16.1.8",
-    "@nguniversal/builders":                  "^16.1.2",
-    "@nx/angular":                            "^16",
-    "@nx/eslint-plugin":                      "^16",
-    "@nx/js":                                 "^16",
-    "@nx/linter":                             "^16",
-    "@nx/workspace":                          "^16",
-    "@schematics/angular":                    "^16.1.6",
-    "@types/node":                            "^18.17.3",
-    "@typescript-eslint/eslint-plugin":       "^5.62.0",
-    "@typescript-eslint/parser":              "^5.62.0",
-    "eslint":                                 "^8.46.0",
-    "eslint-config-prettier":                 "^8.10.0",
-    "eslint-plugin-tsdoc":                    "^0.2.17",
-    "firebase-admin":                         "^11.10.1",
-    "firebase-functions":                     "^4.4.1",
-    "firebase-tools":                         "^12.4.7",
-    "git-describe":                           "^4.1.1",
-    "ng-packagr":                             "^16.1.0",
-    "nx":                                     "^16",
-    "nx-cloud":                               "^16",
-    "postcss":                                "^8.4.27",
-    "postcss-import":                         "^14.1.0",
-    "postcss-preset-env":                     "^7.8.3",
-    "postcss-url":                            "^10.1.3",
-    "prettier":                               "^2.8.8",
-    "typescript":                             "^5.1.6",
-    "url":                                    "^0.11.1"
+    "@angular-eslint/template-parser": "^16.1.0",
+    "@angular/cli": "^16.2.0",
+    "@angular/compiler-cli": "^16.2.0",
+    "@angular/language-service": "^16.2.0",
+    "@nguniversal/builders": "^16.2.0",
+    "@nx/angular": "^16",
+    "@nx/eslint-plugin": "^16",
+    "@nx/js": "^16",
+    "@nx/linter": "^16",
+    "@nx/workspace": "^16",
+    "@schematics/angular": "^16.1.6",
+    "@types/node": "^18.17.5",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint": "^8.47.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-tsdoc": "^0.2.17",
+    "firebase-admin": "^11.10.1",
+    "firebase-functions": "^4.4.1",
+    "firebase-tools": "^12.4.7",
+    "git-describe": "^4.1.1",
+    "ng-packagr": "^16.2.0",
+    "nx": "^16",
+    "nx-cloud": "^16",
+    "postcss": "^8.4.27",
+    "postcss-import": "^14.1.0",
+    "postcss-preset-env": "^7.8.3",
+    "postcss-url": "^10.1.3",
+    "prettier": "^2.8.8",
+    "typescript": "^5.1.6",
+    "url": "^0.11.1"
   }
 }
