@@ -2,7 +2,7 @@ import { Location }                                                             
 import { Injectable, Signal }                                                                                                                                                                                                                                                                               from "@angular/core";
 import { toSignal }                                                                                                                                                                                                                                                                                         from "@angular/core/rxjs-interop";
 import { ActivationEnd, ActivationStart, ChildActivationEnd, ChildActivationStart, GuardsCheckEnd, GuardsCheckStart, NavigationCancel, NavigationEnd, NavigationError, NavigationStart, ResolveEnd, ResolveStart, RouteConfigLoadEnd, RouteConfigLoadStart, Router, RouterEvent, RoutesRecognized, Scroll } from "@angular/router";
-import { filter, map, startWith }                                                                                                                                                                                                                                                                           from "rxjs";
+import { distinctUntilChanged, filter, map, startWith }                                                                                                                                                                                                                                                     from "rxjs";
 
 
 @Injectable({
@@ -18,16 +18,15 @@ export class PathService {
   ) {
     this
       .path$ = toSignal<string>(
-        router.events.pipe<NavigationEnd, string, string>(
+        router.events.pipe<NavigationEnd, string, string, string>(
           filter<RouterEvent | NavigationStart | NavigationEnd | NavigationCancel | NavigationError | RoutesRecognized | GuardsCheckStart | GuardsCheckEnd | RouteConfigLoadStart | RouteConfigLoadEnd | ChildActivationStart | ChildActivationEnd | ActivationStart | ActivationEnd | Scroll | ResolveStart | ResolveEnd, NavigationEnd>(
             (routerEvent: RouterEvent | NavigationStart | NavigationEnd | NavigationCancel | NavigationError | RoutesRecognized | GuardsCheckStart | GuardsCheckEnd | RouteConfigLoadStart | RouteConfigLoadEnd | ChildActivationStart | ChildActivationEnd | ActivationStart | ActivationEnd | Scroll | ResolveStart | ResolveEnd): routerEvent is NavigationEnd => routerEvent instanceof NavigationEnd,
           ),
           map<NavigationEnd, string>(
             (navigationEnd: NavigationEnd): string => navigationEnd.url.split("?")[0],
           ),
-          startWith<string>(
-            location.path(),
-          ),
+          startWith<string, [string]>(location.path()),
+          distinctUntilChanged<string>(),
         ),
         {
           requireSync: true,
