@@ -2,7 +2,7 @@ import { BreakpointObserver, BreakpointState }             from "@angular/cdk/la
 import { DOCUMENT, isPlatformBrowser }                     from "@angular/common";
 import { Inject, Injectable, PLATFORM_ID, signal, Signal } from "@angular/core";
 import { toSignal }                                        from "@angular/core/rxjs-interop";
-import { fromEvent, map, startWith }                       from "rxjs";
+import { distinctUntilChanged, fromEvent, map, startWith } from "rxjs";
 
 
 type ColorScheme = "light" | "dark";
@@ -23,8 +23,9 @@ export class ResponsivityService {
   ) {
     this
       .colorScheme$ = isPlatformBrowser(platformId) ? toSignal<ColorScheme>(
-        breakpointObserver.observe("(prefers-color-scheme: light)").pipe<ColorScheme>(
+        breakpointObserver.observe("(prefers-color-scheme: light)").pipe<ColorScheme, ColorScheme>(
           map<BreakpointState, ColorScheme>((breakpointState: BreakpointState): ColorScheme => breakpointState.matches ? "light" : "dark"),
+          distinctUntilChanged<ColorScheme>(),
         ),
         {
           requireSync: true,
@@ -35,9 +36,10 @@ export class ResponsivityService {
         fromEvent<Event>(
           document,
           "scroll",
-        ).pipe<number, number>(
+        ).pipe<number, number, number>(
           map<Event, number>((): number => document.defaultView?.scrollY || 0),
           startWith<number>(document.defaultView?.scrollY || 0),
+          distinctUntilChanged<number>(),
         ),
         {
           requireSync: true,
